fix(entity/service):图片地址伪哈希索引越界与MySQL字段类型兼容处理

+++ product/product-service/src/main/java/com/kunyiduan/entity/Brand.java
@@ -68,7 +68,7 @@ public class Brand implements Serializable {
     /**
      * 图片地址伪hash索引
      */
-    private Integer picCrc;
+    private Long picCrc;
 
     /**
      * 创建时间
